+++ src/main/java/org/lanternpowered/server/network/vanilla/message/type/play/MessagePlayInOutHeldItemChange.java
@@ -29,23 +29,23 @@ import org.lanternpowered.server.network.message.Message;
 
 public final class MessagePlayInOutHeldItemChange implements Message {
 
-    private final byte slot;
+    private final int slot;
 
     /**
      * Creates a new held item change message.
-     * 
-     * @param slot the new slot
+     *
+     * @param slot The new slot
      */
-    public MessagePlayInOutHeldItemChange(byte slot) {
+    public MessagePlayInOutHeldItemChange(int slot) {
         this.slot = slot;
     }
 
     /**
      * Gets the item slot the player changed to or will change to.
      * 
-     * @return the slot
+     * @return The slot
      */
-    public byte getSlot() {
+    public int getSlot() {
         return this.slot;
     }
